+++ docs/history.md
@@ -1,5 +1,11 @@
 # History
 
+## 1.0.0 (2022-04-07)
+
+* Redis is now an optional dependency
+* Simplify client
+* The Campaign class is now used to create and update a campaign
+
 ## 0.1.0 (2021-06-22)
 
 * First release on PyPI.